Update getWithdrawLiquidityPayload.mdx
In the parameters of the Getwithdrawliquiditypayload method, incorrect links to the Token types in lines are indicated
1)
The reference to the TOKEN type has the wrong format - <(/Types/token)> is a syntactic error. As a result of such links, Markdown will not correctly interpret these links.

The correct format for reference to type in Markdown is to use straight round brackets without corners, as is done in other parts of the documentation. In the documentation by type, the link should look as follows: (/Types/token).

Signed-off-by: savvasmoke <136869149+savvar9991@users.noreply.github.com>